DDF-5258 refactored query for all searches into wrapper
